+++ src/features/profile/components/profile-card/teacher-profile-extra.tsx
@@ -1,16 +1,10 @@
-import { useFormContext } from 'react-hook-form';
-
 import Image from 'next/image';
-
-import { ProfileUpdateForm } from '@/features/profile/schema/schema';
-import { Form, Input } from '@/shared/components/ui';
 
 interface TeacherProfileExtraProps {
   teacherNoteCount: number;
   studentCount: number;
   reviewCount: number;
   description: string;
-  isEditMode: boolean;
 }
 
 export default function TeacherProfileExtra({
@@ -18,13 +12,7 @@ export default function TeacherProfileExtra({
   studentCount,
   reviewCount,
   description,
-  isEditMode,
 }: TeacherProfileExtraProps) {
-  const {
-    register,
-    formState: { errors },
-  } = useFormContext<ProfileUpdateForm>();
-
   return (
     <>
       <div className="flex justify-between">
@@ -74,22 +62,7 @@ export default function TeacherProfileExtra({
 
       <div>
         <h4 className="font-body1-heading mb-2">간단 소개</h4>
-
-        {isEditMode ? (
-          <Form.Item error={!!errors.desc}>
-            <Form.Control>
-              <Input
-                {...register('desc')}
-                placeholder="간단 소개를 입력해 주세요."
-              />
-            </Form.Control>
-            {errors.desc?.message && (
-              <Form.ErrorMessage>{errors.desc.message}</Form.ErrorMessage>
-            )}
-          </Form.Item>
-        ) : (
-          <p>{description}</p>
-        )}
+        <p>{description}</p>
       </div>
     </>
   );
